pyproject.toml
[build-system]
requires = ["uv_build>=0.9.5,<0.10.0"]
build-backend = "uv_build"

[project]
name = "llm"
version = "0.0.3"
description = "A nice python project"
readme = "README.md"
requires-python = ">=3.13"
license = "MIT"
authors = [
    { name = "Mystic" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "fastapi>=0.110.0",
    "matplotlib>=3.10.7",
    "pillow>=12.0.0",
    "prometheus-fastapi-instrumentator>=7.0.0",
    "pydantic-settings>=2.0.0",
    "python-json-logger>=2.0.0",
    "PyYAML>=6.0.3",
    "rich>=14.2.0",
    "seaborn>=0.13.2",
    "tensorboard>=2.20.0",
    "tokenizers>=0.21.0",
    "torch>=2.9.0",
    "transformers>=4.40.0",
    "typer-slim[standard]>=0.20.1",
    "uvicorn>=0.27.0",
]

[dependency-groups]
dev = [
  {include-group = "test"}
]
test = [
    "allure-pytest>=2.15.0",
    "httpx>=0.27.0",
    "pytest>=9.0",
    "pytest-cov>=7.0",
]

[project.urls]
Documentation = "https://github.com/pplmx/llm#readme"
Issues = "https://github.com/pplmx/llm/issues"
Source = "https://github.com/pplmx/llm"

[tool.uv.build-backend]
module-name = "llm"

[tool.pytest]
minversion = "9.0"
filterwarnings = [
    "ignore:The argument 'device' of:DeprecationWarning",
    "ignore:Module llm was previously imported"
]
addopts = [
    "-ra",
    "-q",
]
markers = [
    "quick: Fast unit tests (< 1s)",
    "slow: Slower tests (1-30s)",
    "heavy: Resource-intensive tests (> 30s, may require GPU)",
    "e2e: End-to-end workflow tests",
    "integration: Tests requiring external dependencies (e.g., HuggingFace)",
]
consider_namespace_packages = true # Enable this to allow using the same filename in different test directories without needing __init__.py in each one
pythonpath = ["src", "."]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.run]
source_pkgs = ["llm"]
branch = true
parallel = true
omit = [
    "*/.venv/*",
    "*/tests/*",
    "*/site-packages/*",
]

[tool.coverage.paths]
llm = ["src/llm", "*/llm/src/llm"]
tests = ["tests", "*/llm/tests"]

[tool.coverage.report]
exclude_lines = [
    "if __name__ == .__main__.:",
    "if self.debug:",
    "if TYPE_CHECKING:",
    "no cov",
    "raise NotImplementedError",
    "pass",
    "pragma: no cover",
]
show_missing = true  # Show missing lines in the report
precision = 2  # Number of decimal places to use when rounding
